remove initilize and move :url attribute processing into cas.rb
